pyproject.toml
[project]
name = "ash-rl-mlx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=6.0.0",
    "datasets>=4.5.0",
    "mlx>=0.30.5",
    "mlx-lm>=0.29.1",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "streamlit>=1.53.1",
    "tqdm>=4.67.3",
]
